meams_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pydantic==2.9.2
pydantic[email]
email-validator
python-dotenv==1.0.0
python-multipart==0.0.6
Pillow==10.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
emails==0.6
aiosmtplib==3.0.1
pandas
openpyxl
xlrd
numpy
scipy
statsmodels
python-dateutil
resend